fixin conflicts

+++ Arduino_Programs/SpidyFirmware/comunication.h
@@ -1,5 +1,15 @@
 /**
  * comunication.h
+ * 
+ * Through a call to init_comunication() an i2c comunication will be setup useing the following protocol:
+ *    On receive data request:
+ *      -read number of leg [1..12] or option for send data: 100=ultrasound 101=accelerometer_X 102=accelerometer_Y 103=accelerometer_Z
+ *      -read PWM duty cycle [0..255]
+ *      
+ *    On send data request:
+ *      (requires an option selection from a receive data request)
+ *      -send LSByte from the data asked for
+ *      -send MSByte from the data asked for
  */
 
 #ifndef _COMUNICATION_H_
@@ -12,8 +22,8 @@
   static int state_receive=0;
   static int state_send=0, option;
   static uint8_t legsAngle[12] = {0,0,0,0,0,0,0,0,0,0,0,0};
-  static long duration_U=0, cm_U=0;
-  static long accelerometer_X=0, accelerometer_Y=0, accelerometer_Z=0;
+  static unsigned int duration_U=0, cm_U=0;
+  static int accelerometer_X=0, accelerometer_Y=0, accelerometer_Z=0;
 
   void receiveData(int byteCount);
   void sendData();
